Modularized system service code
